Defined a first iteration of the options design
Decided to go with the builder pattern for easy construction
of options for the user. Underneath, a small hierarchy tree is
used to distinguish between two types of options, namely flagged
an index (this is hidden to user).

Use a strict policy when building objects to prevent the user from
making mistakes, like setting a variable twice.